Split off Connection from HTTPClient.cpp/h
