+++ UWP/CommonUWP/CommonUWP.vcxproj
@@ -99,6 +99,7 @@
     <ClInclude Include="..\..\Common\GPU\Vulkan\VulkanLoader.h" />
     <ClInclude Include="..\..\Common\Math\CrossSIMD.h" />
     <ClInclude Include="..\..\Common\Math\Statistics.h" />
+    <ClInclude Include="..\..\Common\Net\Connection.h" />
     <ClInclude Include="..\..\Common\Net\HTTPNaettRequest.h" />
     <ClInclude Include="..\..\Common\Net\HTTPRequest.h" />
     <ClInclude Include="..\..\Common\Net\NetBuffer.h" />
@@ -285,6 +286,7 @@
     <ClCompile Include="..\..\Common\GPU\GPUBackendCommon.cpp" />
     <ClCompile Include="..\..\Common\GPU\Vulkan\VulkanLoader.cpp" />
     <ClCompile Include="..\..\Common\Math\Statistics.cpp" />
+    <ClCompile Include="..\..\Common\Net\Connection.cpp" />
     <ClCompile Include="..\..\Common\Net\HTTPNaettRequest.cpp" />
     <ClCompile Include="..\..\Common\Net\HTTPRequest.cpp" />
     <ClCompile Include="..\..\Common\Net\NetBuffer.cpp" />
